refactor: split exporter.go into focused files
- exporter_logs.go: struct, lifecycle, and export flow
- clickhouse.go: SQL templates, table constants, batch flush helper
- attributes.go: type conversion, tag/key appending, fingerprinting
